feat: convert CodeBlock to TS
