+++ showcase/app/routes/page-components/code-block.ts
@@ -9,6 +9,11 @@ import {
   ERROR_ICON,
 } from '@hashicorp/design-system-components/components/hds/copy/button/index';
 
+import type { ModelFrom } from 'showcase/utils/ModelFromRoute';
+
+export type PageComponentsCodeBlockModel =
+  ModelFrom<PageComponentsCodeBlockRoute>;
+
 export default class PageComponentsCodeBlockRoute extends Route {
   model() {
     // these are used only for presentation purpose in the showcase
